🚚 Rename Webpack config files
