+++ webpack.config.dev.js
@@ -1,5 +1,5 @@
 const merge = require('webpack-merge');
-const common = require('./webpack.common.js');
+const common = require('./webpack.config.common.js');
 
 module.exports = merge(common, {
   mode: 'development',
